+++ libraries/app/include/graphene/app/database_api.hpp
@@ -43,8 +43,7 @@
 #include <graphene/chain/event_object.hpp>
 #include <graphene/chain/betting_market_object.hpp>
 #include <graphene/chain/global_betting_statistics_object.hpp>
-
-#include <graphene/chain/btc-sidechain/son.hpp>
+#include <graphene/chain/son_object.hpp>
 
 #include <graphene/chain/worker_object.hpp>
 #include <graphene/chain/witness_object.hpp>
