+++ tests/CMakeLists.txt
@@ -45,8 +45,4 @@ file(GLOB RANDOM_SOURCES "random/*.cpp")
 add_executable( random_test ${RANDOM_SOURCES} ${COMMON_SOURCES} )
 target_link_libraries( random_test graphene_chain graphene_app graphene_egenesis_none fc ${PLATFORM_SPECIFIC_LIBS} )
 
-file(GLOB UNIT_TESTS "btc_sidechain_tests/*.cpp")
-add_executable( btc_sidechain_test ${UNIT_TESTS} ${COMMON_SOURCES} )
-target_link_libraries( btc_sidechain_test graphene_chain graphene_app graphene_account_history graphene_bookie graphene_egenesis_none fc graphene_wallet ${PLATFORM_SPECIFIC_LIBS} )
-
 add_subdirectory( generate_empty_blocks )
